Starts migrating JS to TS for components
