+++ app/testimonials/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <p>test</p>;
+}
