+++ components/PaintIcon.jsx
