+++ package.json
@@ -35,12 +35,12 @@
     "swiper": "^9.0.5"
   },
   "devDependencies": {
-    "typescript": "^5",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.0.3",
-    "sass": "^1.57.1"
+    "sass": "^1.57.1",
+    "typescript": "^5"
   }
 }
