+++ src/main.jsx
@@ -1,4 +1,4 @@
-import ReactDOM from "react-dom/client";
+/* import ReactDOM from "react-dom/client";
 import App from "./App";
 import { BrowserRouter } from "react-router-dom";
 
@@ -10,3 +10,4 @@ root.render(
     <App />
   </BrowserRouter>
 );
+ */
